Moved slicer to another file
